Add prettier

+++ contracts/VoteBuy.sol
@@ -25,7 +25,11 @@ contract VoteBuy {
 
     uint256 public blockDelay;
 
-    constructor(address _governor, address _vault, address _dai) {
+    constructor(
+        address _governor,
+        address _vault,
+        address _dai
+    ) {
         governor = GovernorInt(_governor);
         vault = VoteVaultInt(_vault);
         dai = DaiInt(_dai);
@@ -33,7 +37,9 @@ contract VoteBuy {
     }
 
     modifier onlyWhenPoolClosed() {
-        GovernorInt.Proposal memory proposal = governor.proposals(currentProposal);
+        GovernorInt.Proposal memory proposal = governor.proposals(
+            currentProposal
+        );
         require(blockDelay >= (proposal.endBlock - block.number));
         _;
     }
@@ -53,11 +59,11 @@ contract VoteBuy {
         require(dai.transferFrom(msg.sender, address(this), _amount));
 
         // Top up previous buy
-        if(balances[msg.sender].lastBuyProposal == currentProposal) {
+        if (balances[msg.sender].lastBuyProposal == currentProposal) {
             require(balances[msg.sender].aggrement == _aggrement);
             balances[msg.sender].totalBalance += _amount;
             balances[msg.sender].lastBalance += _amount;
-        // New buy (maybe change order to make cheaper for most?)
+            // New buy (maybe change order to make cheaper for most?)
         } else {
             UserStatus memory status = UserStatus({
                 aggrement: _aggrement,
@@ -67,7 +73,7 @@ contract VoteBuy {
             });
             balances[msg.sender] = status;
         }
-        if(_aggrement) {
+        if (_aggrement) {
             currentForPool += _amount;
         } else {
             currentAgainstPool += _amount;
@@ -75,20 +81,23 @@ contract VoteBuy {
     }
 
     function withDraw(address _to) public {
-        if(balances[msg.sender].lastBuyProposal < currentProposal){
+        if (balances[msg.sender].lastBuyProposal < currentProposal) {
             uint256 amount = balances[msg.sender].totalBalance;
             balances[msg.sender].totalBalance = 0;
             dai.transfer(_to, amount);
         } else {
-            uint256 amount = balances[msg.sender].totalBalance - balances[msg.sender].lastBalance;
-            balances[msg.sender].totalBalance = balances[msg.sender].lastBalance;
+            uint256 amount = balances[msg.sender].totalBalance -
+                balances[msg.sender].lastBalance;
+            balances[msg.sender].totalBalance = balances[msg.sender]
+                .lastBalance;
             dai.transfer(_to, amount);
         }
     }
 
     function topUpWithTotalBalance(bool _aggrement) public {
-        if(balances[msg.sender].lastBuyProposal == currentProposal) {
-            balances[msg.sender].lastBalance = balances[msg.sender].totalBalance;
+        if (balances[msg.sender].lastBuyProposal == currentProposal) {
+            balances[msg.sender].lastBalance = balances[msg.sender]
+                .totalBalance;
         } else {
             UserStatus memory status = UserStatus({
                 aggrement: _aggrement,
@@ -98,7 +107,6 @@ contract VoteBuy {
             });
             balances[msg.sender] = status;
         }
-       
     }
 }
 
@@ -118,15 +126,20 @@ interface GovernorInt {
         bool canceled;
         bool executed;
     }
-    function proposals(uint256 _proposalId) external returns(Proposal memory);
+
+    function proposals(uint256 _proposalId) external returns (Proposal memory);
 }
 
 interface VoteVaultInt {
-   
     function vote(uint256 _proposalId, bool _support) external;
 }
 
 interface DaiInt {
-    function transferFrom(address src, address dst, uint rawAmount) external returns (bool);
+    function transferFrom(
+        address src,
+        address dst,
+        uint256 rawAmount
+    ) external returns (bool);
+
     function transfer(address dst, uint256 rawAmount) external returns (bool);
 }
